# Migrate bench to a cobra style command