--- a/cmd/bbolt/command_root.go
+++ b/cmd/bbolt/command_root.go
@@ -20,6 +20,7 @@ func NewRootCommand() *cobra.Command {
 		newVersionCobraCommand(),
 		newSurgeryCobraCommand(),
 		newInspectCobraCommand(),
+		newBenchCobraCommand(),
 	)
 
 	return rootCmd