--- a/tests/robustness/powerfailure_test.go
+++ b/tests/robustness/powerfailure_test.go
@@ -90,11 +90,11 @@ func doPowerFailure(t *testing.T, du time.Duration, fsMountOpt string, useFailpo
 	dbPath := filepath.Join(root, "boltdb")
 
 	args := []string{"bbolt", "bench",
-		"-work", // keep the database
-		"-path", dbPath,
-		"-count=1000000000",
-		"-batch-size=5", // separate total count into multiple truncation
-		"-value-size=512",
+		"--work", // keep the database
+		"--path", dbPath,
+		"--count=1000000000",
+		"--batch-size=5", // separate total count into multiple truncation
+		"--value-size=512",
 	}
 
 	logPath := filepath.Join(t.TempDir(), fmt.Sprintf("%s.log", t.Name()))