@ivanvc 2024-03-27
@ahrtr, I deleted the `-read-mode` flag, as the only option available was `seq`. Providing any other value halted the execution.

However, changing from built-in Go flags to Cobra flags isn't backward compatible, as Go flags aren't POSIX compliant.